requirements.txt
numpy
torch
scikit-learn
matplotlib
pyarrow
pillow
dedalus>=3.0.0
h5py
scipy
mpi4py